ecommerce/requirements.txt
asgiref==3.6.0
Brotli==1.0.9
cssselect2==0.7.0
Django==3.2
django-filter==21.1
django-jazzmin==2.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==5.2.2
Faker==12.0.1
fonttools==4.39.0
gunicorn==20.0.4
html5lib==1.1
mixer==7.2.2
Pillow==9.4.0
pycparser==2.21
pydyf==0.5.0
PyJWT==2.6.0
pyphen==0.13.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tinycss2==1.2.1
uritemplate==4.1.1
weasyprint==58.1
webencodings==0.5.1
zopfli==0.2.2
